+++ src/jv.mli
@@ -387,7 +387,7 @@ module Error : sig
   | `No_modification_allowed_error | `Not_allowed_error | `Not_found_error
   | `Not_readable_error | `Not_supported_error | `Operation_error
   | `Quota_exceeded_error | `Read_only_error | `Security_error
-  | `Syntax_error | `Timeout_error | `Transaction_inactive_error
+  | `Syntax_error | `Timeout_error | `Transaction_inactive_error | `Type_error
   | `Type_mismatch_error | `Url_mismatch_error | `Unknown_error
   | `Version_error | `Wrong_document_error
   | `Other (** If not listed. {b Do not match on this !} *) ]